Add initial implementation of large file mode for TokenizedBuffer
When passed largeFileMode: true, the TokenizedBuffer spawns a background
task to build the initial set of tokenized lines, so that expensive
computation related to scanning for leading whitespace, hard tabs,
paired characters, etc is performed in the background.

This means the TokenizedBuffer won’t be ready to use synchronously
after construction, and we’ll need to handle this in the DisplayBuffer,
TextEditor, and the view as well.